split to template classes
