+++ tests/stairlight/test_config.py
@@ -2,7 +2,8 @@ import os
 from collections import OrderedDict
 
 import src.stairlight.config as config
-import src.stairlight.template as template
+import src.stairlight.source.base as base
+from src.stairlight.source.fs import FsTemplate
 
 
 class TestSuccess:
@@ -31,9 +32,9 @@ class TestSuccess:
         assert list(stairlight_template.keys()) == ["include", "exclude", "settings"]
 
     def test_build_mapping_template(self):
-        sql_template = template.SQLTemplate(
+        sql_template = FsTemplate(
             mapping_config=self.configurator.read(prefix=config.MAPPING_CONFIG_PREFIX),
-            source_type=template.SourceType.FS,
+            source_type=base.TemplateSourceType.FS,
             file_path="tests/sql/main/test_undefined.sql",
         )
         unmapped = [
@@ -74,5 +75,5 @@ class TestSuccess:
         )
 
         expected = {"mapping": [mapping_value], "metadata": [metadata_value]}
-        actual = self.configurator.build_mapping_template(unmapped=unmapped)
+        actual = self.configurator.build_mapping_template(unmapped_templates=unmapped)
         assert actual == expected
